Add files via upload
In recent years, India has become one of the most exciting places for start-ups to operate and grow. From its large population to the sheer number of financial opportunities, India offers considerable potential for entrepreneurs. The country’s supportive environment for startups has drawn in a significant number of investors, both domestic and foreign, resulting in a significant influx of funding into the startup ecosystem.